validators: Moved subtopic validators back to prod_validators to satisfy testing code coverage
