--- core/domain/subtopic_validators.py
@@ -1,190 +1,0 @@
-# coding: utf-8
-#
-# Copyright 2020 The Oppia Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS-IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Validators for subtopic prod models."""
-
-from __future__ import absolute_import  # pylint: disable=import-only-modules
-from __future__ import unicode_literals  # pylint: disable=import-only-modules
-
-from core.domain import base_model_validators
-from core.domain import subtopic_page_domain
-from core.domain import subtopic_page_services
-from core.platform import models
-import python_utils
-
-(
-    base_models, topic_models, user_models, subtopic_models
-) = models.Registry.import_models([
-    models.NAMES.base_model, models.NAMES.topic,
-    models.NAMES.user, models.NAMES.subtopic
-])
-
-
-class SubtopicPageModelValidator(base_model_validators.BaseModelValidator):
-    """Class for validating SubtopicPageModel."""
-
-    @classmethod
-    def _get_model_id_regex(cls, item):
-        return '^%s-\\d*$' % (item.topic_id)
-
-    @classmethod
-    def _get_model_domain_object_instance(cls, item):
-        return subtopic_page_services.get_subtopic_page_from_model(item)
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        snapshot_model_ids = [
-            '%s-%d' % (item.id, version) for version in python_utils.RANGE(
-                1, item.version + 1)]
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'subtopic_page_commit_log_entry_ids',
-                subtopic_models.SubtopicPageCommitLogEntryModel,
-                ['subtopicpage-%s-%s'
-                 % (item.id, version) for version in python_utils.RANGE(
-                     1, item.version + 1)]),
-            base_model_validators.ExternalModelFetcherDetails(
-                'snapshot_metadata_ids',
-                subtopic_models.SubtopicPageSnapshotMetadataModel,
-                snapshot_model_ids),
-            base_model_validators.ExternalModelFetcherDetails(
-                'snapshot_content_ids',
-                subtopic_models.SubtopicPageSnapshotContentModel,
-                snapshot_model_ids),
-            base_model_validators.ExternalModelFetcherDetails(
-                'topic_ids', topic_models.TopicModel, [item.topic_id])]
-
-    @classmethod
-    def _get_custom_validation_functions(cls):
-        return []
-
-
-class SubtopicPageModelValidator(base_model_validators.BaseModelValidator):
-    """Class for validating SubtopicPageModel."""
-
-    @classmethod
-    def _get_model_id_regex(cls, item):
-        return '^%s-\\d*$' % (item.topic_id)
-
-    @classmethod
-    def _get_model_domain_object_instance(cls, item):
-        return subtopic_page_services.get_subtopic_page_from_model(item)
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        snapshot_model_ids = [
-            '%s-%d' % (item.id, version) for version in python_utils.RANGE(
-                1, item.version + 1)]
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'subtopic_page_commit_log_entry_ids',
-                subtopic_models.SubtopicPageCommitLogEntryModel,
-                ['subtopicpage-%s-%s'
-                 % (item.id, version) for version in python_utils.RANGE(
-                     1, item.version + 1)]),
-            base_model_validators.ExternalModelFetcherDetails(
-                'snapshot_metadata_ids',
-                subtopic_models.SubtopicPageSnapshotMetadataModel,
-                snapshot_model_ids),
-            base_model_validators.ExternalModelFetcherDetails(
-                'snapshot_content_ids',
-                subtopic_models.SubtopicPageSnapshotContentModel,
-                snapshot_model_ids),
-            base_model_validators.ExternalModelFetcherDetails(
-                'topic_ids', topic_models.TopicModel, [item.topic_id])]
-
-    @classmethod
-    def _get_custom_validation_functions(cls):
-        return []
-
-
-class SubtopicPageSnapshotMetadataModelValidator(
-        base_model_validators.BaseSnapshotMetadataModelValidator):
-    """Class for validating SubtopicPageSnapshotMetadataModel."""
-
-    EXTERNAL_MODEL_NAME = 'subtopic page'
-
-    @classmethod
-    def _get_model_id_regex(cls, unused_item):
-        return '^[A-Za-z0-9]{1,%s}-\\d*-\\d*$' % base_models.ID_LENGTH
-
-    @classmethod
-    def _get_change_domain_class(cls, unused_item):
-        return subtopic_page_domain.SubtopicPageChange
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'subtopic_page_ids',
-                subtopic_models.SubtopicPageModel,
-                [item.id[:item.id.rfind(base_models.VERSION_DELIMITER)]]),
-            base_model_validators.ExternalModelFetcherDetails(
-                'committer_ids', user_models.UserSettingsModel,
-                [item.committer_id])]
-
-
-class SubtopicPageSnapshotContentModelValidator(
-        base_model_validators.BaseSnapshotContentModelValidator):
-    """Class for validating SubtopicPageSnapshotContentModel."""
-
-    EXTERNAL_MODEL_NAME = 'subtopic page'
-
-    @classmethod
-    def _get_model_id_regex(cls, unused_item):
-        return '^[A-Za-z0-9]{1,%s}-\\d*-\\d*$' % base_models.ID_LENGTH
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'subtopic_page_ids',
-                subtopic_models.SubtopicPageModel,
-                [item.id[:item.id.rfind(base_models.VERSION_DELIMITER)]])]
-
-
-class SubtopicPageCommitLogEntryModelValidator(
-        base_model_validators.BaseCommitLogEntryModelValidator):
-    """Class for validating SubtopicPageCommitLogEntryModel."""
-
-    EXTERNAL_MODEL_NAME = 'subtopic page'
-
-    @classmethod
-    def _get_model_id_regex(cls, item):
-        # Valid id: [subtopicpage]-[subtopic_id]-[subtopic_version].
-        regex_string = '^(subtopicpage)-%s-\\d*$' % (
-            item.subtopic_page_id)
-
-        return regex_string
-
-    @classmethod
-    def _get_change_domain_class(cls, item):
-        if item.id.startswith('subtopicpage'):
-            return subtopic_page_domain.SubtopicPageChange
-        else:
-            cls._add_error(
-                'model %s' % base_model_validators.ERROR_CATEGORY_ID_CHECK,
-                'Entity id %s: Entity id does not match regex pattern' % (
-                    item.id))
-            return None
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'subtopic_page_ids',
-                subtopic_models.SubtopicPageModel,
-                [item.subtopic_page_id])]
